ai_chatbot_RAG-main/requirements.txt
sentence-transformers==2.2.2
rank-bm25==0.2.2
PyPDF2==3.0.1
groq==0.4.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
flask==3.0.0
